requirements.txt
agentfield
pydantic>=2.0
python-dotenv
sentence-transformers>=2.2.0
numpy>=1.24.0
tree-sitter>=0.21.0
watchfiles>=0.21.0
